source/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astroid==2.15.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
autopep8==1.6.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
databases==0.6.2
dill==0.3.6
docstring-to-markdown==0.12
ecdsa==0.18.0
fastapi==0.85.0
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
idna==3.4
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
ormar==0.12.1
parso==0.8.3
passlib==1.7.4
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0rc1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pydocstyle==6.2.3
pyflakes==3.0.1
pylint==2.17.1
python-dotenv==1.0.0
python-jose==3.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
PyYAML==6.0
requests==2.28.2
rope==1.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.25.11
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
whatthepatch==1.0.4
wrapt==1.15.0
yapf==0.32.0
yarl==1.8.2
backoff==2.2.1
rocketmq-client-python==2.0.0
redis==4.5.4
fastapi_pagination==0.12.3